backend/requirements.txt
fastapi==0.115.11
python-dotenv==1.0.1
uvicorn==0.34.0
upstash-redis==1.4.0
firebase-admin==6.7.0
sqlalchemy
psycopg2
pydantic